Warning removed.
git-svn-id: svn://svn.code.sf.net/p/hxcfloppyemu/code/HxCFloppyEmulator/libhxcfe/trunk@679 63fa2f70-6a9b-4bc4-b855-9c6a949b1d69

+++ libhxcfe/sources/loaders/raw_loader/raw_writer.c
@@ -113,7 +113,7 @@ int RAW_libWrite_DiskFile(HXCFLOPPYEMULATOR* floppycontext,FLOPPY * floppy,char 
 							{
 								if(sca[k]->sector==l)
 								{
-									if(sca[k]->sectorsize!=sectorsize)
+									if(sca[k]->sectorsize!=(unsigned int)sectorsize)
 									{
 										sectorsize=-1;
 									}
